fix for bug 5632
git-svn-id: https://svn.code.sf.net/p/specify/code/trunk/Specify@4358 ce9bf4bf-376f-1848-a496-a1c9fe3b1a39

+++ src/edu/ku/brc/specify/config/SpecifyAppContextMgr.java
@@ -1718,13 +1718,16 @@ public class SpecifyAppContextMgr extends AppContextMgr
             {
                 session = DataProviderFactory.getInstance().createSession();
                 session.beginTransaction();
-                appResDir.getSpPersistedAppResources().remove(appRes);
-                appResDir.getSpAppResources().remove(appRes);
-                session.saveOrUpdate(appResDir);
+                if (!appResDir.removeResource(appRes))
+                {
+                    session.rollback();
+                    log.error("Unable to remove AppResource '" + appResource + "' from directory '" + appResDirName + "'");
+                    return false;
+                }
                 session.delete(appRes);
+                session.saveOrUpdate(appResDir);
                 session.commit();
                 session.flush();
-                
                 return true;
             } catch (Exception ex)
             {
